requirements.txt
fastapi[all]
SQLalchemy
pydantic-settings
python-jose
bcrypt
passlib
psycopg[binary]
alembic
pytest